Introduce GetFeedPage usecase
