+++ app/src/main/java/com/olderwold/reddit/MainViewModel.kt
@@ -1,17 +1,11 @@
 package com.olderwold.reddit
 
 import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
 import androidx.paging.Pager
 import androidx.paging.PagingConfig
-import androidx.paging.PagingSource
-import com.olderwold.reddit.data.RedditClient
 import com.olderwold.reddit.domain.FeedItem
 import com.olderwold.reddit.ui.RedditPagingSource
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
