+++ app/src/main/java/com/olderwold/reddit/domain/GetFeedPage.kt
@@ -1,0 +1,8 @@
+package com.olderwold.reddit.domain
+
+interface GetFeedPage {
+    suspend operator fun invoke(
+        limit: Int,
+        after: String?
+    ): FeedPage
+}
