+++ app/src/test/java/com/olderwold/reddit/ui/RedditPagingSourceTest.kt
@@ -1,7 +1,7 @@
 package com.olderwold.reddit.ui
 
 import androidx.paging.PagingSource
-import com.olderwold.reddit.data.RedditClient
+import com.olderwold.reddit.data.NetworkGetFeedPage
 import com.olderwold.reddit.data.dto.RedditHot
 import com.olderwold.reddit.domain.FeedItem
 import com.olderwold.reddit.tape
@@ -18,7 +18,7 @@ class RedditPagingSourceTest {
     @Test
     @OkReplay(tape = "smoke_test")
     fun `load should succeed with loading the data from the tape`(): Unit = runBlocking {
-        val redditClient = RedditClient(testRule.api, RedditHot.Mapper())
+        val redditClient = NetworkGetFeedPage(testRule.api, RedditHot.Mapper())
         val pagingSource = RedditPagingSource(redditClient)
 
         val loadParams = PagingSource.LoadParams.Refresh<String>(
